Fix Port When using Webpack
**Bug**
I'm using this package with webpack, on a test page at `localhost:4000`. This ends up making requests to api.giphy.com:4000/... which always fail. It seems that the Webpack shim for `http`/`https` fills in the port using `window.location.port` if it is not explicitly provided.

**Fix**
Change request options to use `host` instead of `hostname`. This stops webpack setting the port from `window.location.port`.